Remove unused function from core
